Fixed #45.
