+++ test/files/pos/i45.scala
@@ -1,0 +1,7 @@
+trait DisjunctionFunctions
+trait MaybeFunctions {
+  def empty[A]: Option[A]      = None
+}
+
+trait AllFunctions extends DisjunctionFunctions with MaybeFunctions
+object Scalaz extends AllFunctions
